Bump `google-protobuf` to 3.25.3
This was removed from the Gemfile.lock during [the Ruby 3.3
update](https://github.com/alphagov/asset-manager/commit/90209aa81deb9b58f77441207ed1bb34decd99a2)
but not removed as an actual dependency, therefore any subsequent runs
of `bundle install` have resulted in the incompatible version of the gem
being reinstalled.

Therefore upgrading to a version of `google-protobuf` that is compatible
with Ruby 3.3.