Ensure asset metadata is updated when file name isn't changed
Asset metadata is not being updated when an asset is uploaded with a file which
has the same name as the previous file. This is because of a condition added in
#1258 that was intended to quieten some errors we were seeing at the time. This
same issue was addressed in #1263.

This is problematic as there is a guard statement in the S3 uploader that
prevents assets with an unchanged MD5 hexdigest from uploading.