Trivial: standardise on netcat-openbsd.
Shouldn't matter much but might as well keep this consistent (both with
what we used to install under 22.04 and with what's in
govuk-infrastructure/toolbox).

The OpenBSD implementation lacks `-e` (which is a plus) and is a bit
more modern and actively maintained.

netcat is used in the healthcheck script in the underlying clamav
container:
https://github.com/Cisco-Talos/clamav-docker/blob/main/clamav/1.3/debian/scripts/clamdcheck.sh