Fix race condition where file is deleted while updating assets
The SaveToCloudStorageWorker is kicked off as soon as virus scanning is complete and the asset transitions to the clean state.

When the SaveToCloudStorageWorker completes the upload, it triggers the upload_success event, which transitions to the uploaded state.

There's an after_transition callback which runs after the asset moves into the uploaded state, which removes the uploaded file.

Meanwhile, there's a before_save callback which will update the file metadata before saving the asset. This callback will only run if the file hasn't been uploaded.

A race condition occurs if the before_save callback has started when the asset transitions into the uploaded state - one process deletes the file, while the other process continues to perform methods which expect the file to exist.

We can avoid this by delaying the deletion of the file. Once the asset is in the uploaded state, we won't perform operations which expect the file to exist in the before_save callback. There's no urgency behind removing the file, so we can do it a few minutes after the state transition.