Reuse AWS client between S3 requests
At the moment, we create new AWS clients whenever we download or head an object. Each time we do this, the client requests a new set of AWS credentials using the pod's (or the node's?) metadata endpoint.

This causes a problem when asset-manager is under high load, because the AWS metadata endpoint has a rate limit. Eventually, we start getting 429 errors back from the metadata endpoint, which the AWS client reports as:

    Aws::Sigv4::Errors::MissingCredentialsError: missing credentials

(At least according to issue #2823 in aws-sdk-ruby)

The S3Storage instance in asset-manager is created once when the app loads, and set as Services.cloud_storage. In other words it's effectively a singleton - the same instance will be used by all requests.

Aws::S3::Client instances are thread safe, and can be reused between requests (again, according to issue #2823 in aws-sdk-ruby).

By creating the client once, instead of once per request, we'll avoid hitting the AWS metadata endpoint on every request to S3, which should mean that we don't get rate limited.