Validate public assets have public document URLs
There are over 1000 assets in the Asset Manager database which are not drafts, but have a parent document URL that points to the draft stack.

We expect this may help us to diagnose an issue where at least some of those 1000+ assets were unexpectedly made public ahead of their parent document being published. If this validation had been in place we think it would have prevented the issue.