Do not attempt to open file if it does not exist
`File.stat(path)` raises `Errno::ENOENT` if the file does not exist.

In the case that a file has already been deleted (e.g. by another
worker), we don't need to raise an error, just fail gracefully.

Therefore only attempting to load the file if it actually exists.

Also adds a test that covers the behaviour that caught this issue:
attemting to soft-delete a file that has already been removed from the
filesystem.