Fix invalid state and reset parent document url when updating assets
- Add a clause to remove parent document url when fixing draft states when deleting superseded assets
Some assets are incorrectly staying in the draft state. But when we try and fix the draft state, it cannot be linking to a draft parent_document_url. We are removing the parent document url link since it is difficult to figure out what the parent_document_url should be, and is largely irrelevant for a superseded asset.
- Fix correct state for assets that have gone in a weird invalid state over the years.
"deleted" should no longer be a valid state. patching any data that we come across that has a wrong state.
- Log when we've patched an invalid state or parent document URL