Allow updates to draft deleted assets
In publishing apps, we have the following issue:

1. User replaces an attachment asset with a new one on a new edition of a published document.
2. User deletes the attachment.
3. User publishes the new edition of the document.
4. Replaced attachment asset remains live.

This happens because the deleted asset still has a value of "false" for the draft attribute,
so the replacement on the old asset is not active.

In order to address this, we are allowing the update of draft deleted assets.
Given that assets are still live on GOV.UK until they are both "deleted" and "live"
(i.e. have a draft value of "false"), we are more accurately fulfilling the intent
of the API by allowing updates to deleted draft assets.