Accept some changes suggested by Rails 8 upgrade
After upgrading to Rails 8, running `rails app:update` suggests a load
of changes to framework files based on the defaults that a brand new
Rails 8 app will come with out-of-the-box.

It can be useful to adopt these changes to try to keep up with the
standard behaviour that Rails developers are building on top of and our
own developers will come to expect Rails apps in general to exhibit.

For Asset Manager I've followed these general rules:

* Not introducing new static files because I'm not sure that this app
  has any use for them?
* Accepting suggested changes to existing static files because I don't
  have the time to understand how/if they're already useful
* Accept new default config that's only replacing old default config and
  that either doesn't seem all that significant anyway or that I can't
  tell whether it's useful or not without spending too much time on
* Skip the rest