Return '410 Gone' status code for deleted assets (WHIT-2301)
We get semi-regular requests from publishers to manually apply
redirects to old 404'ing assets, because those assets are still
surfacing in Google search results. Applying redirects eventually
causes Google to surface the redirected page instead, and drop
the old 404'ing asset from its results.

Patching in a redirect on a per-case basis is not scalable nor
sustainable. The underlying issue is that publishers expect
assets to stop appearing in search when they're deleted.

Historically, Google has generally kept 404 responses in its
index for longer (treating the response code as a temporary blip),
whereas when it encounters a 410 response, it immediately removes
the page from its index. There are various Wikipedia and SEO blog
posts that substantiate that, but I'm probably best linking to the
[MDN docs](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Status/410):
> The HTTP 410 Gone client error response status code indicates that the target resource is no longer available at the origin server and that this condition is likely to be permanent. A 410 response is cacheable by default.
> Clients should not repeat requests for resources that return a 410 response, and website owners should remove or replace links that return this code. If server owners don't know whether this condition is temporary or permanent, a 404 status code should be used instead.

As of July 2025, Google does seem to treat 404s and 410
[interchangeably](https://developers.google.com/search/docs/crawling-indexing/http-network-errors#pagegone),
so this may not have the remove-from-index effect that we'd hope.
But if they're following the HTTP spec, in theory it should lead
to faster removals from the index, and in any case, returning a
410 response is a more descriptive and useful response than
generalising everything as a 404.

JIRA: WHIT-2301
Zendesk: #6169516