Handle all missing and deleted assets in ApplicationController
Code review suggestion:
https://github.com/alphagov/asset-manager/pull/1734/files#r2209681059

Having rescues spread across several controllers was confusing.
It was necessary because we were raising different exception types
depending on the 'flavour' of asset (Whitehall or Media). This
commit consolidates onto one type of exception for all, making it
much easier and cleaner to handle capturing those exceptions at
the ApplicationController level.

The continued presence of
`rescue_from Mongoid::Errors::DocumentNotFound, with: :error_404`
is a bit annoying, but a number of tests break if we remove it,
which suggests there are places in the code that can trigger
that exception outside of the controller workflows that would
otherwise return our custom `MediaErrors::AssetNotFound`.