Do not raise an error if another replacement is invalid
In the `update_indirect_replacements_on_publish` and
`backpropagate_replacement` callback methods, we were using `save!` to
update previous replacements. This results in an error being raised if
the other replacement is invalid.

When data is invalid, we do not want to raise an error but instead
report the error to the user. Therefore updating to use `update` instead
of `save!`, to prevent these being logged.

The change from `save` to `save!` (and therefore the change in
behaviour) was added in ba0bd1c14c60c41b38ef2a8298e52622c45363fd after
enforcing a rubocop rule. The change in error handling did not appear to
be intentional.